Renamed spotify.generated.interface to spotify-http.interface
